Handle case where plotting no data
Had the following error today:
   gnuplot> plot for [i=0:*] "last-week-inside.txt" index i using 2:3 with lines title word(titles,i+1)
                                                                                                       ^
            line 0: x range is invalid

Issue was that last-week-inside.txt was empty because there was no
inside measurement taken for the week at the time of generating the
graph, and gnuplot then exit with a non-zero status.

Keep the "set -e" on the script, we're still expecting the command to
succeed.